[TASKSCHED] Add support for tasks with time specification, and differentiate between them and "simple" timeless tasks.
We support a structured text "task list file" with one task per line:
   time <whitespace> Task_description
where:
- time is in hour:minutes (HH:MM) format. This is optional.
- Task_description is a one-line string describing the task.
The tasks' timestamp is parsed with std::get_time().

Internally, tasks are represented as CTask class objects, and are stored
in two different queues, as they are treated in different orders:
"timeless", i.e. "simple" enumerated tasks, versus timed tasks.
--> One single queue wouldn't be sufficient.

Timed tasks get sorted so as to be popped out in ascending order.
--> This requires comparison operator(s) for CTasks to be implemented.

NOTE: Since I try to stick to the original aim of the exercise, by using
std::queue<>, I have to introduce that manual sorting function to sort
the queue, and whose implementation is quite inefficient (to simplify it
I move all the queued elements to an std::vector<> so as to be able to
simply call std::sort() on it -- not possible for std:queue<> --, and
then move the sorted elements back into the queue).
It would have been better to use a more sophisticated queue-like
structure instead, that supports sorting natively...

For ease of usage, operator<< is overloaded for printing out task details
to output streams.

This code illustrates:
- Structures / classes: https://youtu.be/iVLQeWbgbXs etc.
- Templates: https://youtu.be/qrJjFN4Igfw

and other concepts:
- time_t, tm;
- Overloading comparison and injection operators.